Back-end/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
boilerpy3==1.0.7
cachetools==5.5.2
cattrs==24.1.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
construct==2.5.3
distro==1.9.0
docopt==0.6.2
Events==0.5
faiss-cpu==1.11.0
farm-haystack==1.26.4.post0
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
haystack==0.42
httpcore==1.0.8
httpx==0.27.2
huggingface-hub==0.30.2
idna==3.10
inflect==7.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy-imports==0.3.1
lxml==5.4.0
MarkupSafe==3.0.2
monotonic==1.6
more-itertools==10.7.0
mpmath==1.3.0
narwhals==1.36.0
networkx==3.4.2
num2words==0.5.14
numpy==2.2.5
ollama==0.3.3
packaging==24.2
pandas==2.2.3
pefile==2024.8.26
pillow==11.2.1
platformdirs==4.3.7
posthog==3.25.0
prompthub-py==4.0.0
protobuf==5.29.4
pyarrow==19.0.1
pydantic==1.10.21
pydantic_core==2.33.1
pydeck==0.9.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.20
python-ptrace==0.9.9
pytz==2025.2
PyYAML==6.0.2
quantulum3==0.9.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-cache==0.9.8
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==79.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sseclient-py==1.8.0
starlette==0.46.2
streamlit==1.44.1
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
typeguard==4.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
url-normalize==2.2.0
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
Werkzeug==3.1.3
